WorkflowClient, ScheduleClient and ScheduleHandle: added methods `withTimeout()`, `withDeadline()` and `withRetryOptions()`

+++ src/Client/GRPC/Connection/ConnectionInterface.php
@@ -9,12 +9,14 @@ interface ConnectionInterface
     public function getState(bool $tryToConnect = false): ConnectionState;
 
     /**
-     * @param int<0, max> $timeout in microseconds
+     * Wait for the channel to be ready.
+     *
+     * @param float $timeout in seconds
      *
      * @return bool true if channel is ready
      * @throws \Exception if channel is in FATAL_ERROR state
      */
-    public function waitForReady(int $timeout): bool;
+    public function waitForReady(float $timeout): bool;
 
     public function close(): void;
 }
